Bull Cow game demo
A demo I put together based on code originally from an Unreal
programming class on Udemy.com. The idea is to guess a word based on how
many letter exist in the word in your guess. Letters that are in the
right position in the word are bulls, and letter that exist in the guess
but are not in the right position are cows. For example if the word is
planes and I guessed plants I would have 5 bulls.